+++ resources/views/components/form/partials/string.blade.php
@@ -7,6 +7,7 @@
     'autofocus' => false,
     'autocomplete' => 'off',
     'transliterate' => false,
+    'dir' => config('text_dir'),
 ])
 
 @php
@@ -14,7 +15,7 @@
 @endphp
 
 @if ($disabled)
-    <input type="{{ $type }}" id="{{ $id }}" value="{{ $value }}" disabled {!! $attributes->merge(['class' => 'input text disabled'])->toHtml() !!} autocomplete="off" />
+    <input type="{{ $type }}" id="{{ $id }}" value="{{ $value }}" dir="{{ $dir }}" disabled {!! $attributes->merge(['class' => 'input text disabled'])->toHtml() !!} autocomplete="off" />
     <input type="hidden" name="{{ $name }}" value="{{ $value }}" />
 @else
     <input type="{{ $type }}"
@@ -22,6 +23,7 @@
         id="{{ $id }}"
         value="{{ $value }}"
         autocomplete="{{ $autocomplete }}"
+        dir="{{ $dir }}"
         @if ($autofocus) autofocus @endif
         @if ($transliterate)
             data-transliterate-code="{{ $transliterate }}"