+++ resources/views/components/log-data.blade.php
@@ -4,7 +4,7 @@
             <td colspan="2">
                 <div class="table table-log-head">
                     <div class="cell cell-log-head left">
-                        <span class="text-big">{{ $key }}</span>
+                        <span class="text-big" dir="ltr">{{ $key }}</span>
                     </div>
                     <div class="cell cell-log-head right">
                         <x-cms::form.button type="button" bg="icons.download" class="btn-icon btn-normal js-list-button" data-type="download" data-filename="{{ get_relative_path(storage_path('logs/'.$key)) }}" title="{{ __('main.title_download') }}"/>
@@ -21,7 +21,7 @@
                 </td>
                 <td class="log-message">
                     <div class="log-message-hold">
-                        <span class="text">{!! soft_break_string($entry['message']) !!}</span>
+                        <span class="text" dir="ltr">{!! soft_break_string($entry['message']) !!}</span>
                     </div>
                 </td>
             </tr>