+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ##### Fixed
 - Exception catch in redirect check
+- Redundant redirect in admin
 
 #### [2.4.0] - 2025-11-20